[pnnl/SHAD#63] implemented and tested all numeric ops + refactoring
